chore: move token row to its own file
